Day01 Part03 - End of Part03
This is the first of several zipped files that represent the end of certain sections in the course. These files correspond to the Microsoft Word document Day01 ReactJS.docx. In order to re-create the original app, you must first extract these files into a directory. Open a terminal window and navigate into the folder where the files have been extracted. Run the command "npm install". That command will then download all the supporting files needed for this application to work properly.